==> transcriptome/step3-enrichment.R <==
rm(list = ls())   
options(stringsAsFactors = F)


load(file = 'step2-DEG-macular.Rdata')

head(nrDEG)


logFC_cutoff=log2(1.5)
nrDEG$change = as.factor(ifelse(nrDEG$pvalue  < 0.05 & abs(nrDEG$log2FoldChange) > logFC_cutoff,
                              ifelse(nrDEG$log2FoldChange > logFC_cutoff ,'UP','DOWN'),'NOT'))


table(nrDEG$change)
head(nrDEG)
sub_nrDEG = nrDEG[nrDEG$change %in% c("UP","DOWN"),]
sub_nrDEG$ENSEMBL=rownames(sub_nrDEG)
library(ggplot2)
library(clusterProfiler)
library(org.Hs.eg.db)
library(stringr)
df <- bitr(unique(sub_nrDEG$ENSEMBL), fromType = "ENSEMBL",
           toType = c( "SYMBOL","ENTREZID"),
           OrgDb = org.Hs.eg.db)
head(df)
nrDEG=sub_nrDEG
head(nrDEG)

nrDEG=merge(nrDEG,df,by='ENSEMBL')

head(nrDEG)
table(nrDEG$change)
write.csv(nrDEG,file = "anno_nrDEG_macular.csv")
save(nrDEG,file = 'anno_nrDEG_macular.Rdata')


gene_up= nrDEG[nrDEG$change == 'UP','ENTREZID'] 
gene_down=nrDEG[nrDEG$change == 'DOWN','ENTREZID'] 
gene_diff=c(gene_up,gene_down)
# gene_all=as.character(nrDEG[ ,'ENTREZID'] ) 

## KEGG pathway analysis
### 做KEGG数据集超几何分布检验分析，重点在结果的可视化及生物学意义的理解。
# 包装的函数
{
  run_kegg <- function(gene_up,gene_down,geneList=F,pro='test'){
    gene_up=unique(gene_up)
    gene_down=unique(gene_down)
    gene_diff=unique(c(gene_up,gene_down))
    ###   over-representation test
    # 下面把3个基因集分开做超几何分布检验
    # 首先是上调基因集。
    kk.up <- enrichKEGG(gene         = gene_up,
                        organism     = 'hsa',
                        #universe     = gene_all,
                        pvalueCutoff = 0.05,
                        qvalueCutoff =0.05)
    head(kk.up)[,1:6]
    kk=kk.up
    pdf("KK_UP_barplot.pdf",useDingbats = F,width = 8.5,height = 8.5)
    dotplot(kk)
    dev.off()
    kk=DOSE::setReadable(kk, OrgDb='org.Hs.eg.db',keyType='ENTREZID')
    write.csv(kk@result,paste0(pro,'_kk.up.csv'))
    
    # 首先是下调基因集。
    kk.down <- enrichKEGG(gene         =  gene_down,
                          organism     = 'hsa',
                          #universe     = gene_all,
                          pvalueCutoff = 0.9,
                          qvalueCutoff =0.9)
    head(kk.down)[,1:6]
    kk=kk.down
    dotplot(kk)
    kk=DOSE::setReadable(kk, OrgDb='org.Hs.eg.db',keyType='ENTREZID')
    write.csv(kk@result,paste0(pro,'_kk.down.csv'))
    
    # 最后是上下调合并后的基因集。
    kk.diff <- enrichKEGG(gene         = gene_diff,
                          organism     = 'hsa',
                          pvalueCutoff = 0.9,
                          qvalueCutoff =0.9)
    head(kk.diff)[,1:6]
    kk=kk.diff
    dotplot(kk)
    kk=DOSE::setReadable(kk, OrgDb='org.Hs.eg.db',keyType='ENTREZID')
    write.csv(kk@result,paste0(pro,'_kk.diff.csv'))
    
    
    kegg_diff_dt <- as.data.frame(kk.diff)
    kegg_down_dt <- as.data.frame(kk.down)
    kegg_up_dt <- as.data.frame(kk.up)
    down_kegg<-kegg_down_dt[kegg_down_dt$p.adjust<0.05,];down_kegg$group=-1
    up_kegg<-kegg_up_dt[kegg_up_dt$p.adjust<0.05,];up_kegg$group=1
    #画图设置, 这个图很丑，大家可以自行修改。
    g_kegg=kegg_plot(up_kegg,down_kegg)
    print(g_kegg)
    
    ggsave(g_kegg,filename = paste0(pro,'_kegg_up_down.png') )
    
    if(geneList){
      ###  GSEA 
      ## GSEA算法跟上面的使用差异基因集做超几何分布检验不一样。
      kk_gse <- gseKEGG(geneList     = geneList,
                        organism     = 'hsa',
                        nPerm        = 1000,
                        minGSSize    = 20,
                        pvalueCutoff = 0.9,
                        verbose      = FALSE)
      head(kk_gse)[,1:6]
      gseaplot(kk_gse, geneSetID = rownames(kk_gse[1,]))
      gseaplot(kk_gse, 'hsa04110',title = 'Cell cycle') 
      kk=DOSE::setReadable(kk_gse, OrgDb='org.Hs.eg.db',keyType=keyType,)
      tmp=kk@result
      write.csv(kk@result,paste0(pro,'_kegg.gsea.csv'))
      
      
      # 这里找不到显著下调的通路，可以选择调整阈值，或者其它。
      down_kegg<-kk_gse[kk_gse$pvalue<0.05 & kk_gse$enrichmentScore < 0,];down_kegg$group=-1
      up_kegg<-kk_gse[kk_gse$pvalue<0.05 & kk_gse$enrichmentScore > 0,];up_kegg$group=1
      
      g_kegg=kegg_plot(up_kegg,down_kegg)
      print(g_kegg)
      ggsave(g_kegg,filename = paste0(pro,'_kegg_gsea.png'))
      # 
    }
    
  }
  
  ### GO database analysis 
  ### 做GO数据集超几何分布检验分析，重点在结果的可视化及生物学意义的理解。
  run_go <- function(gene_up,gene_down,pro='test'){
    gene_up=unique(gene_up)
    gene_down=unique(gene_down)
    gene_diff=unique(c(gene_up,gene_down))
    g_list=list(gene_up=gene_up,
                gene_down=gene_down,
                gene_diff=gene_diff)
    # 因为go数据库非常多基因集，所以运行速度会很慢。
    if(T){
      go_enrich_results <- lapply( g_list , function(gene) {
        lapply( c('BP','MF','CC') , function(ont) {
          cat(paste('Now process ',ont ))
          ego <- enrichGO(gene          = gene,
                          #universe      = gene_all,
                          OrgDb         = org.Hs.eg.db,
                          ont           = ont ,
                          pAdjustMethod = "BH",
                          pvalueCutoff  = 0.99,
                          qvalueCutoff  = 0.99,
                          readable      = TRUE)
          
          print( head(ego) )
          return(ego)
        })
      })
      save(go_enrich_results,file =paste0(pro, '_go_enrich_results.Rdata'))
      
    }
    # 只有第一次需要运行，就保存结果哈，下次需要探索结果，就载入即可。
    load(file=paste0(pro, '_go_enrich_results.Rdata'))
    
    n1= c('gene_up','gene_down','gene_diff')
    n2= c('BP','MF','CC') 
    for (i in 1:3){
      for (j in 1:3){
        fn=paste0(pro, '_dotplot_',n1[i],'_',n2[j],'.png')
        cat(paste0(fn,'\n'))
        png(fn,res=150,width = 1080)
        print( dotplot(go_enrich_results[[i]][[j]] ))
        dev.off()
      }
    }
    
    
  }
  
  kegg_plot <- function(up_kegg,down_kegg){
    dat=rbind(up_kegg,down_kegg)
    colnames(dat)
    dat$pvalue = -log10(dat$pvalue)
    dat$pvalue=dat$pvalue*dat$group 
    
    dat=dat[order(dat$pvalue,decreasing = F),]
    
    g_kegg<- ggplot(dat, aes(x=reorder(Description,order(pvalue, decreasing = F)), y=pvalue, fill=group)) + 
      geom_bar(stat="identity") + 
      scale_fill_gradient(low="blue",high="red",guide = FALSE) + 
      scale_x_discrete(name ="Pathway names") +
      scale_y_continuous(name ="log10P-value") +
      coord_flip() + theme_bw()+theme(plot.title = element_text(hjust = 0.5))+
      ggtitle("Pathway Enrichment") 
  }
  
  
}
run_kegg(gene_up,gene_down,pro='kegg_macular_')



go <- enrichGO(gene_up, OrgDb = "org.Hs.eg.db", ont="all") 
library(ggplot2)
library(stringr)
barplot(go, split="ONTOLOGY")+ facet_grid(ONTOLOGY~., scale="free")
barplot(go, split="ONTOLOGY",font.size =10)+ 
  facet_grid(ONTOLOGY~., scale="free") + 
  scale_x_discrete(labels=function(x) str_wrap(x, width=50))+
  ggsave("go_UP_ALL_barplot.pdf",useDingbats = F,width = 11,height = 8.5) 
go=DOSE::setReadable(go, OrgDb='org.Hs.eg.db',keyType='ENTREZID')
write.csv(go@result,'macular_gene_up_GO_all.csv')


go <- enrichGO(gene_down, OrgDb = "org.Hs.eg.db", ont="all") 
barplot(go, split="ONTOLOGY",font.size =10)+ 
  facet_grid(ONTOLOGY~., scale="free") + 
  scale_x_discrete(labels=function(x) str_wrap(x, width=50))+
  ggsave("go_down_ALL_barplot.pdf",useDingbats = F,width = 11,height = 8.5)
go=DOSE::setReadable(go, OrgDb='org.Hs.eg.db',keyType='ENTREZID')
write.csv(go@result,'macular_gene_down_GO_all.csv')

go <- enrichGO(gene_diff, OrgDb = "org.Hs.eg.db", ont="all") 
barplot(go, split="ONTOLOGY",font.size =10)+ 
  facet_grid(ONTOLOGY~., scale="free") + 
  scale_x_discrete(labels=function(x) str_wrap(x, width=50))+
  ggsave("go_diff_ALL_barplot.pdf",useDingbats = F,width = 11,height = 8.5)
go=DOSE::setReadable(go, OrgDb='org.Hs.eg.db',keyType='ENTREZID')
write.csv(go@result,'macular_gene_diff_GO_all.csv')
